% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkshift.R
\name{bulkshift}
\alias{bulkshift}
\title{Sonar backscatter bulk shift}
\usage{
bulkshift(
  shift,
  target,
  preds = NULL,
  model = "glm",
  mosaic = FALSE,
  mosaicmethod = "bilinear",
  savedata = TRUE,
  sample = NULL,
  samplemethods = "uniform",
  crossvalidate = NULL,
  ...
)
}
\arguments{
\item{shift}{SpatRaster. Backscatter dataset undergoing correction.}

\item{target}{SpatRaster. Backscatter dataset used as reference.}

\item{preds}{SpatRaster. One or more layers to use as additional predictor variables for backscatter calibration model.}

\item{model}{character. Method used to model error between backscatter datasets. Currently supported models are "mean" (i.e., intercept-only), "glm", and "randomForest". Model defaults are retained; use "..." to specify additional parameters.}

\item{mosaic}{logical. Mosaic and output the backscatter layers?}

\item{mosaicmethod}{character. Which method used to resample the corrected backscatter layer for mosaicking? See \code{\link[terra:resample]{terra::resample()}} for details.}

\item{savedata}{logical. Whether to output the model data.frame.}

\item{sample}{numeric. Proportion of overlapping data to sample for modelling using \link{samplemethods.}}

\item{samplemethods}{character vector. The method used to subsample the data. One or several of "uniform" (the default), "stratify", or "autocorrelation". Specifying multiple methods results in a combined output. See \link{Details}.}

\item{crossvalidate}{numeric. Proportion of data to use for validation. Validation data are drawn from the dataset following saubsampling if \link{sample} is used.}

\item{...}{Additional parameters to pass to models or subsampling methods.}
}
\description{
Relative calibration of two partially overlapping sonar backscatter datasets using the methods presented in Misiuk et al. (2020). This is largely a wrapper for the \link{terra} package to facilitate relative calibration of backscatter rasters. Supports several modelling methods and optionally, spatially explicit subsampling and validation.
}
\references{
Misiuk, B., Brown, C.J., Robert, K., Lacharite, M., 2020. Harmonizing Multi-Source Sonar Backscatter Datasets for Seabed Mapping Using Bulk Shift Approaches. Remote Sensing 12, 601. https://doi.org/10.3390/rs12040601
}
