% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{Overlap}
\usage{
overlap(x, y)
}
\arguments{
\item{x}{SpatRaster. Backscatter dataset undergoing correction.}

\item{y}{SpatRaster. Will be algined and/or projected to x if extent, origin, or crs do not match.}
}
\value{
SpatRaster
}
\description{
Extract the union of overlapping SpatRasters with different extents, origins, and projections.
}
