% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bError.R
\name{bError}
\alias{bError}
\title{Backscatter error visualization}
\usage{
bError(x, interactive = TRUE, cex = 0.75, pch = 16, alpha = 1)
}
\arguments{
\item{x}{An (unmodified) list output from \link{bulkshift}.}

\item{interactive}{Logical. If additional predictors were supplited to \link{bulkshift}, should an interactive 3D plot be returned?}

\item{cex}{Size of scatterplot points.}

\item{pch}{Scatterplot point symbols.}

\item{alpha}{Alpha (transparency) of points in scatterplot.}
}
\description{
Visualize the error between backscatter datasets, and optionally, additional predictors, in 2 and 3D.
}
\details{
This function generates 2D or 3D scatterplots from the output of \link{bulkshift} using the \link{plot3D} and \link{plot3Drgl} packages to assist in diagnosing the quality of model fit.

The output shows the error between backscatter datasets as a function of the "shift" dataset on the x-axis.
If no other predictors were used for \link{bulkshift}, this is a 2D scatterplot.
If additional predictors were supplied, they are plotted on the z-axis and a 3D scatterplot is generated.
}
\examples{
bb2016 <- rast(system.file('extdata', 'bb2016.tif', package='bulkshift'))
bb2017 <- rast(system.file('extdata', 'bb2017.tif', package='bulkshift'))
bbdepth <- rast(system.file('extdata', 'bbdepth.tif', package='bulkshift'))

#2D example
b <- bulkshift(bb2017, bb2016)
bError(b)

#3D example
b <- bulkshift(bb2017, bb2016, bbdepth)
bError(b)

}
