% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bExplore.R
\name{bExplore}
\alias{bExplore}
\title{Exploratory backscatter layer analysis}
\usage{
bExplore(
  x,
  y,
  preds = NULL,
  error_map = TRUE,
  boxplot = TRUE,
  error_plot = TRUE,
  loess = TRUE,
  loess_samp = 10000,
  ...
)
}
\arguments{
\item{x, y}{\link{terra} SpatRasters. Backscatter layers to compare.}

\item{preds}{\link{terra} SpatRaster. Predictors to explore for explaining error between \code{x} and \code{y}.}

\item{error_map}{Logical. Whether to plot a map of error between \code{x} and \code{y}.}

\item{boxplot}{Logical. Whether to draw boxplots comparing \code{x} and \code{y}. See Details.}

\item{error_plot}{Logical. Whether to plot the error as a function of \code{x}, and also \code{preds} if specified.}

\item{loess}{Logical. Whether to use \link{loess} to visualized the relationship between \code{preds} (including \code{x}) and the error.}

\item{loess_samp}{Numeric. Maximum number of samples used to estimate the \link{loess} smoother. See Details.}

\item{...}{Additional arguments to pass to \link{loess}. See Details.}
}
\description{
Explore backscatter layer statistics and relationships prior to harmonization to inform the bulk shift procedure.
}
\details{
This function creates a series of helpful plots for comparing backscatter datasets prior to attempting harmonization (e.g., using \link{bulkshift}).
Hit enter to proceed through the plots.

\code{boxplot = TRUE} draws box plots comparing the full distributions of x and y, and also the distributions where they overlap. This
is useful for assessing the representativeness of the area overlap (i.e., whether the error model will need to extrapolate).

Local polynomial fitting using loess can be slow with large sample sizes. Subsampling using \code{loess_samp} is used here only to fit
the smoother; it is predicted on the entire dataset. Additional arguments can be passed to \code{...} (see \link{loess} for more information).
}
\examples{
bb2016 <- rast(system.file('extdata', 'bb2016.tif', package='bulkshift'))
bb2017 <- rast(system.file('extdata', 'bb2017.tif', package='bulkshift'))
bbdepth <- rast(system.file('extdata', 'bbdepth.tif', package='bulkshift'))

bExplore(bb2017, bb2016, preds = bbdepth)

}
