% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSample.R
\name{bSample}
\alias{bSample}
\title{Advanced subsampling of a SpatRaster for spatial modelling}
\usage{
bSample(x, size, samplemethods = c("uniform"))
}
\arguments{
\item{x}{SpatRaster layer}

\item{size}{sample size}

\item{samplemethods}{vector of subsampling methods. One or multiple of "uniform", "stratify", "autocorrelation".}
}
\value{
Vector of cell numbers
}
\description{
Advanced subsampling of a SpatRaster for spatial modelling
}
