% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSample.R
\name{bSample}
\alias{bSample}
\title{Subsampling methods for a SpatRaster}
\usage{
bSample(x, size, samplemethods = c("uniform"))
}
\arguments{
\item{x}{SpatRaster layer}

\item{size}{Sample size.}

\item{samplemethods}{Vector of subsampling methods. One or multiple of "uniform", "stratify", "autocorrelation".}
}
\value{
Vector of cell numbers
}
\description{
Advanced subsampling of a \link{terra} SpatRaster for spatial modelling
}
\details{
If \code{samplemethods = "stratify"}, samples are distributed equally among quartiles of x.
\code{samplemethods = "autocorrelation"} weights sampling by the inverse of local Moran's I, which is
calculated used \code{\link[terra:autocor]{terra::autocor()}}.
}
