% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkshift.R
\name{bulkshift}
\alias{bulkshift}
\title{Sonar backscatter bulk shift}
\usage{
bulkshift(
  shift,
  target,
  preds = NULL,
  model = "glm",
  mosaic = FALSE,
  mosaicmethod = "bilinear",
  savedata = TRUE,
  sample = NULL,
  samplemethods = "uniform",
  crossvalidate = NULL,
  ...
)
}
\arguments{
\item{shift}{SpatRaster. Backscatter dataset undergoing correction.}

\item{target}{SpatRaster. Backscatter dataset used as reference.}

\item{preds}{SpatRaster. One or more layers to use as additional predictor variables for backscatter calibration model.}

\item{model}{Character. Method used to model error between backscatter datasets. Currently supported models are "mean" (i.e., intercept-only), "glm", and "randomForest". Model defaults are retained; use \code{...} to specify additional parameters.}

\item{mosaic}{Logical. Mosaic and output the backscatter layers?}

\item{mosaicmethod}{Character. Which method used to resample the corrected backscatter layer for mosaicking? See \code{\link[terra:resample]{terra::resample()}} for details.}

\item{savedata}{Logical. Whether to output the model data.frame. If \code{crossvalidate} is used, the validation data is additionally returned as \code{dataVal}.}

\item{sample}{Numeric. Proportion of overlapping data to sample for modelling using samplemethods.}

\item{samplemethods}{Character vector. The method used to subsample the data. One or several of "uniform" (the default), "stratify", or "autocorrelation". Specifying multiple methods results in a combined output. See Details.}

\item{crossvalidate}{Numeric. Proportion of data to use for validation. Validation data are drawn from the dataset following saubsampling if \code{sample} is used.}

\item{...}{Additional parameters passed to the model.}
}
\value{
List of bulkshift objects
}
\description{
Relative calibration of two partially overlapping sonar backscatter datasets using the methods presented in Misiuk et al. (2020). This is largely a wrapper for the \link{terra} package to facilitate relative calibration of backscatter rasters. Supports several modelling methods and optionally, spatially explicit subsampling and validation.
}
\details{
By default, this function fits a linear regression to the error between two backscatter
datasets using \link{glm}. The modelling method can be changed using the \code{model} argument, but extrapolation
is important, and regression has performed well in simulations. Additional predictors such as bathymetry
can be added as a single SpatRaster (potentially with multiple layers) using \code{preds}.

\code{sample} calls on \link{bSample} to draw a random sample of the data for modelling. This has two purposes:
\enumerate{
\item to facilitate computation for large datasets, and
\item two represent the data more equitably (e.g., where backscatter values are over-represented at a certain level).
}

To handle the latter issue, \code{samplemethods} can be set to "stratify", which uses backscatter data quartiles to stratify the sampling.
Setting \code{samplemethods} to "autocorrelation" weights the sampling spatially using local Moran's I,
as calculated by \code{\link[terra:autocor]{terra::autocor()}}, so that sampling is reduced at areas of high autocorrelation.
Multiple \code{samplemethods} can be combined by providing them as a vector.
}
\examples{
bb2016 <- rast(system.file('extdata', 'bb2016.tif', package='bulkshift'))
bb2017 <- rast(system.file('extdata', 'bb2017.tif', package='bulkshift'))
bbdepth <- rast(system.file('extdata', 'bbdepth.tif', package='bulkshift'))

#run bulkshift using bathymetry as an additional predictor
b <- bulkshift(bb2017, bb2016, bbdepth)
plot(b$shifted)

#cross-validate the model and create a mosaic
b <- bulkshift(bb2017, bb2016, bbdepth, mosaic = TRUE, crossvalidate = 0.25)
b$fitStats
b$testStats

}
\references{
Misiuk, B., Brown, C.J., Robert, K., Lacharite, M., 2020. Harmonizing Multi-Source Sonar Backscatter Datasets for Seabed Mapping Using Bulk Shift Approaches. Remote Sensing 12, 601. https://doi.org/10.3390/rs12040601
}
